services/recommender/requirements.txt
# FastAPI и связанные зависимости
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pydantic>=1.8.0

# OpenAI клиент для работы с LLM
openai>=1.0.0

# HTTP клиент для взаимодействия с другими микросервисами
requests>=2.25.0

# Для работы с переменными окружения
python-dotenv>=0.19.0

# Для разработки и тестирования
pytest>=6.2.0
pytest-asyncio>=0.15.0
